+++ scglobals/static/tx/src/app/helpers/pageNumbers.js
@@ -5,6 +5,7 @@ module.exports = function(skip,limit,total) {
   var start = showDecimals((Number(skip) + 1));
   var end = (Number(skip) + Number(limit));
   end = (end > total) ? total : end;
+  debugger;
   end = showDecimals(end);
   return start + " - " + end;
 };