+++ sqlite_testing.sql
@@ -4,19 +4,19 @@
 -- .mode column 
 
 -- all stock changes
--- select s.id,count(sc.id)
+-- select s.id,count(sc.id) as countsc
 --     from inventory_shipment s
 --     join inventory_stockchange sc on sc.shipment_id=s.id
 --     where s.active=1 
---     -- and count(sc.id) > 0
+--     and coucountsc > 0
 --     group by s.id
 --     limit 1000000;
 -- all stock changes
-select s.id from inventory_shipment s 
-    join inventory_stockchange sc on sc.shipment_id=s.id 
-    where s.active=1 
-    -- and sc.qty > 0
-    group by s.id;
+-- select s.id from inventory_shipment s 
+--     join inventory_stockchange sc on sc.shipment_id=s.id 
+--     where s.active=1 
+--     and sc.qty = 0;
+--     group by s.id;
 
 -- select 
 --   s.date,s.id,sc_location.name,from_location.name,from_location.location_type,to_location.name,to_location.location_type,i.name,c.name,il.expiration,il.lot_num,il.unit_price,sc.qty,u.username,sc.modified,(il.unit_price*sc.qty)